+++ scripts/dts/extract/__init__.py
@@ -1,0 +1,7 @@
+#
+# Copyright (c) 2017 Bobby Noelte
+#
+# SPDX-License-Identifier: Apache-2.0
+#
+
+# Empty to allow all modules to be imported
